Conversation UX: one write-capable conversation, unified zoom, remove trash icon
- Collapse Ask/Edit into a single write-capable Conversation entry
  (confirm-before-write via the read-write system prompt)
- Remove trash/new-conversation button from the conversation menu bar;
  deletes happen via the outline context menu
- Unify text resizing: conversation.zoom + WKWebView.pageZoom on the
  transcript, and sync the composer font on update (latent bug fix)
- Remove dead QueryConversationView; relocate QueryMode +
  showsEditingBanner into ConversationView
- Clean up stale 'Ask tab' preflight errors
