+++ Tests/WikiFSTests/QueryModeTests.swift
@@ -56,25 +56,25 @@ struct QueryModeTests {
 
     /// Ask mode, not generating → no banner.
     @Test func bannerPredicate_ask_notGenerating() {
-        #expect(QueryConversationView.showsEditingBanner(
+        #expect(ConversationView.showsEditingBanner(
             allowsEdits: QueryMode.ask.allowsEdits, isGenerating: false) == false)
     }
 
     /// Ask mode, generating → no banner (Ask is read-only regardless).
     @Test func bannerPredicate_ask_generating() {
-        #expect(QueryConversationView.showsEditingBanner(
+        #expect(ConversationView.showsEditingBanner(
             allowsEdits: QueryMode.ask.allowsEdits, isGenerating: true) == false)
     }
 
     /// Edit mode, not generating → no banner (agent idle).
     @Test func bannerPredicate_edit_notGenerating() {
-        #expect(QueryConversationView.showsEditingBanner(
+        #expect(ConversationView.showsEditingBanner(
             allowsEdits: QueryMode.edit.allowsEdits, isGenerating: false) == false)
     }
 
     /// Edit mode, generating → banner shown (agent CAN write, ingestion is paused).
     @Test func bannerPredicate_edit_generating() {
-        #expect(QueryConversationView.showsEditingBanner(
+        #expect(ConversationView.showsEditingBanner(
             allowsEdits: QueryMode.edit.allowsEdits, isGenerating: true) == true)
     }
 }
